Remove db credentials from public repo.
