+++ .gitignore
@@ -21,3 +21,6 @@ vendor/
 .env
 .docker_build/
 bin/
+
+# DB login credentials
+db.go
